sia-backend/server/requirements.txt
Flask==3.0.3;
uWSGI==2.0.27;
flask_favicon==0.1.4;
psycopg2-binary==2.9.10;
